+++ scripts/add.tokens.erc20.mjs
@@ -6,7 +6,8 @@ import { readFileSync, writeFileSync } from 'node:fs';
 import { resolve } from 'node:path';
 import path from 'path';
 import { ENV } from './build.utils.mjs';
-import { CK_ERC20_JSON_FILE } from './utils.mjs';
+
+import { CK_ERC20_JSON_FILE } from './constants.mjs';
 
 dotenv.config({ path: `.env.${ENV}` });
